Add infrastructure for periodic keepalive probes
Use AlarmManager to wake up the phone every <n> minutes and generate
traffic to keep the connection up.  <n> is user-selectable; the default
value is 14 (half of the default Cisco ASA idle timeout, minus 1 minute).

This doesn't actually generate the traffic yet, but it does set up the
alarm upon connection, and cancel it upon disconnection.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>