VPNProfileList: Prompt for hostname, not profile name, when adding VPN
For an AnyConnect VPN, the hostname is the only mandatory field.  So
make it easy for users to enter the minimal information necessary to
connect to their gateway.

Now the profile name will default to the hostname, but the user can
change them both individually from the settings editor.

Also, add code to gray out the OK button until at least one character
has been entered, to help avoid creating a profile with an empty hostname.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>