VPNProfileList: Use URI text entry consistently for hostname
Currently, editing an existing hostname in a VPN profile uses URI entry,
but adding a new VPN profile uses TYPE_TEXT_FLAG_NO_SUGGESTIONS only.
This results in different keys being available on the touch keyboard,
often replacing the '.com' key with a smiley emoticon in the latter case.

So we will change the input type used in the Add operation to use URI
entry.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>