KeepAlive: Fix socket binding across wifi/3G handoffs
On Android 4.4+, the DNS request socket can get bound to a specific
upstream interface instead of the VPN interface.  After a handoff,
sending data traffic to this socket can return EINVAL and log a Java
exception.  Fix this by creating and tearing down the socket every time
we try to send a DNS keepalive.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>