Make "About OpenConnect" a menu item, not a tab
Most cell phones do not have a wide enough screen to display all five of
our tabs in portrait mode, which means that one tab will be invisible
until the user scrolls over and finds it.  This is rather crummy from
a usability standpoint, so let's sacrifice the "about" tab and move it
into its own activity (accessed through the menu).

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>