external: Build curl and libstoken using openconnect's standalone toolchain
Since openconnect now installs its own standalone toolchains and populates
the sysroot with libraries we can use, it eliminates the need to build
extra copies of libtomcrypt and openssl for stoken and curl.

To some extent this limits parallelism in the build, but it also results
in compiling fewer modules.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>