AuthFormHandler: Fix "rotation accepts input" problem
In BATCH_MODE_EMPTY_ONLY, the check for empty fields is performed every
time the dialog is redrawn.  This means that if the screen is rotated
or if the app switches between foreground->background->foreground and
there is partial form input, the partial entry will be accepted when the
form is redrawn.

The fix is to remember the allFilled state for the lifetime of the auth
form.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>