OpenConnectManagementThread: Fix base64 stored file detection
Android's base64 decoder rejects "some" bad inputs, but not all of them.
This means that if a VPN profile had stored a literal certificate instead
of a base64-encoded certificate (as was the case in <= 0.7), it might get
decoded into junk data.  Through experimentation and reading the source
code, it was shown that inputs like "a&bc" or "ab*c" do not raise
exceptions.

So we will add a regexp to reject any base64 string with invalid
characters.  At the minimum this should cause PEM files starting with
"-----BEGIN " to be processed as literal values, not as base64.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>