TokenImportActivity: Process sdtid files
Add a file association with *.sdtid and allow them to be imported manually
from the VPN profile editor.  These may require a password but never
require entering a devid.

As is the case with passworded CTF strings, the token will be converted
into an unpassworded CTF string and stored in the profile.  We will not
store the whole XML file.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>