OpenConnectManagementThread: Fix regression on CSD fixup
Commit a543c06cf428 (Allow use of x86 Linux CSD wrapper scripts) added
logic to check for e.g. "#!/bin/sh" at the start of the script, and
prepend "#!/system/bin/sh" if the interpreter was missing.  In a
subsequent commit, d08793cfcb (Copy certs into individual files), this
logic was inadvertently bypassed; stored CSD scripts would be used as-is
instead of rewritten to add the corrected interpreter line.

So, for any stored file for which isExecutable == true, we will copy
the contents to a temporary file and change the interpreter if necessary.
Handling of actual certs (not CSD scripts) should be unaffected.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>