OpenConnectManagementThread: Revert back to old SHA1 cert hash
Commit 948a60430ff5e0 ("Use new Java API to get peer cert") inadvertently
changed the format of the certificate hash, which would cause previously
accepted certs to generate a new warning.  The new scheme only hashes a
portion of the cert; the traditional cert fingerprint computes the SHA1
of the entire DER file.

So for now we will manually implement the latter scheme and preserve
compatibility with existing stored hashes.

See also:
http://thread.gmane.org/gmane.network.vpn.openconnect.devel/1956/focus=2113

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>